@@ -2,10 +2,14 @@ package com.ashomok.lullabies.di_dagger;
 
 import android.app.Application;
 import android.content.Context;
+import android.content.SharedPreferences;
 
+
+import com.ashomok.lullabies.R;
 
 import dagger.Binds;
 import dagger.Module;
+import dagger.Provides;
 
 /**
  * This is a Dagger module. We use this to bind our Application class as a Context in the AppComponent
@@ -18,11 +22,11 @@ import dagger.Module;
  */
 @Module
 public abstract class ApplicationModule {
-//    @Provides
-//    static SharedPreferences provideSharedPrefs(Context context) {
-//        return context.getSharedPreferences(
-//                context.getString(R.string.preferences), Context.MODE_PRIVATE);
-//    }
+    @Provides
+    static SharedPreferences provideSharedPrefs(Context context) {
+        return context.getSharedPreferences(
+                context.getString(R.string.preferences), Context.MODE_PRIVATE);
+    }
 
     //expose Application as an injectable context
     @Binds
